// src/app/features/home/domain/dto/tmdb-credits.dto.ts
export interface TmdbCastDto {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface TmdbCrewDto {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  credit_id: string;
  department: string;
  job: string;
}

export interface TmdbCreditsDto {
  id: number;
  cast: TmdbCastDto[];
  crew: TmdbCrewDto[];
}